Uploading Previously completed day1 Challenge.
Fairly easy, just checking specific sum from a list of strings that need to be converted to it.
Hard coded because it was faster than implementing the proper number of iterations and so on.

Took longer than expected because I was coding without visual studio, but via Repl.it, and also needed to remind myself how to use StreamReader to get the lines in text file.

Part 1: 12 mins
Part 2: 2 mins