Pushing previously solved day2 challenge.
Started challenge on time at 05:00.

Checking if a string (password) contains a boundry range of a given character, for solution 1.
Took time to figure out how to quickly count the number of chars in a string, no simple method built in for that, so had to use Linq, which not familiar in... (probably would have spent less time if just made a method that goes through all chars in string and increases char count if reads a certain char... oh well)

Solution 2 was easier, due to need to have a char XOR at 2 given positions.

Took longer than wanted because I didn't know quick way to iterate through the string to get char count of a specific char. Also during solution 2, forgot the indexes given in input start at 1, so had to -1 the index, but instead I thought it was part of the challenge so implemented a whole try catch block for it... What a donkey.

Part 1: 30 mins
Part 2: 9 mins