Pushing day3 challenge completed today
Started late at 05:25

Part 1: 41 mins
Part 2: 8 mins

A fun challenge! (made a visual Print of the map being traversed at it is insane how big it gets in some of the sled runs in solution 2)

For Solution 1, had to count the number of a char #, when traversing a map, as if travelled in a diagonal line, e.g. by down row 1, right column 3.

The interesting part of the challenge is that the input given is a repeating unit of the map. So when the sled travels outside the boundry of the input, the input repeats, so I just used try catch, to detect when sled is outside of the map, and then made the sled go back, and appended the repeating unit to the map again.

For solution 2, needed to test other sled runs, with different given gradient (moving down rows and right columns). So just added travelRight & travelDown ints to TraverseMap method, so can work with different gradients. And added list of sleds that held all sled runs.
The sneaky buggers decided to make the multiplication to get the answer a very large number, which a normal int in c# would be turned negative, so had to spend extra time using a long instead of int.

Solution 1: problems
Took me a long time to figure out the best way to store and access the map that will be traversed, wrote unneccessay methods to get the char[] from any given map...
Also spent some time looking up how to catch multiple exceptions at the same time, because either row accessed or column accessed would both fail at different times, so had to look up catching both exceptions, because I didn't want to copy paste same catch exception block, with just a different exception.

Solution 2 problems:
Hard coded the sleds initially, which was not useful when a negative int was outputed after mutliplication, so wasted time un hard coding using a list, which took additional time.