Pushing day4 challenge.
Started on time at 05:00

Part 1: 39.5 mins
Part 2: 44.5 mins

Not such a fan of this challenge. Was just data processing and parsing.

For Solution 1, had to sort the lines of input files into an individual's fields.
The challenge was seperate each individual, and each field within the individual, wrote 2 seperate methods, then thought i didn't know each individual, when in fact I had a List<List<string>>, where first List<> was individual...
On second thought I should have used a Dictionary to store the key:value pairs, instead of parsing seperately...
After seperating the individuals' fields, had to check if all correct fields are present to be counted as valid entry.

For Solution 2, had to actually validate the attributes given to the field.
It was long because there was 8 different cases of validation, 3 of which are similar, 3 last ones were easy, but the other 2 were trickier.
I think there are definetly predefined methods I could have used to validate, getting a number and a string from a single string, or convert a char to hex digit, but I could't find any in System.Numerics, so decided to make myself.