Pushing day5 challenge
Started late at 6:23

Part1: 38 mins
Part2: 1 hour 12 mins

Part 1 was interesting, learning binary space partinioning.
Part 2 was unclear for me, so spent such a long time doing it.

For part 1 of the challenge you have to seperate the a line into 2 fields, rows and columns. Then using binary space partitioning, to figure out a number from a series of lower/upper chars from both rows and columns. Using each one's row and column figure out a seat ID, return the highest seat ID.

Part 2, I didn't think it was very clear at all. Just sort of said some seats are missing in front and back, and that you can +1 or -1 to get to your actual seat id. So tried figuring out where it could be and wasted time guessing spaces as if it was another binary space partioning exercsise.
When I got no results after trying 3 times, I found the missing entries, and took away missing num entries from my entry, it was not the answer...
Then found all the missing entries with a lower id than highest ID, and took them away. Still was not the answer....
So then printed out all the missing ids, and found a strange missing id in the middle of the continuous missing front and back numbers. So hard inputed the answer... And guess what! It was the answer....