Pushing day9
Started late at 05:15

Part1: 55mins
Part2: 35 mins

Good challenge, an interesting data validation challenge, and learning contiguos summation.

For part1, had to seperate the input into "preamble" nums and "validation" nums. Each validation num must be made up by sum of 2 preamble nums, the smae preamble nums can be used again, but not their sum value.
However, the plot twist that cost me some time was, the preamble foreach validation num changes, because the preamble is always the 25 nums before it, so previous validation nums, can be used as preamble, and previous preamble nums have to be removed.
Also another plot twist was the big numbers, which made my int overflow, so had to make it long.
Had to find the the first validation number which doesn't meet this criteria.

For part2, had to find sum of the min and max value, within the contiguous sum of the answer to part1.  Iterate through each number within all the nums in the input,  following another iteration through the numbers, but with a varying starting index, which is set ot the first loop counter. Summing all nums within second loop, if found is equal to the part1 answer, then give the indexes of start of the second loop, and the current index within the second loop. This acts as a range for which numbers can be used to find min/max values.
Then simply find minimum and maximum values within that range, and sum them.