Pushing day11 Challenge
Started late at 07:36

Part1:  1 hour 7 mins
Part2: After a long time, after coming back from college

Be warned, this code is not pretty....

It was  a tricky challenge.
Main problems with part2 was that i have forgotten that checking each GridPoint within the grid as a potential neighbour, would cause not surrounding tiles to be checked, so would give a false answer.
Also, forgot the grid travelling downwards actually means the y index increasing, and when going upwards, the y index decreases.... Idiot