Pushing day12
Started challenge late at 05:20

Part1: 1 hour 3 mins
Part2: 45 mins

A good cahllenge, which i tackled well, and with good code.
Actually seperating the code into classes, to make it easier to manage.

Part1 problem, when getting direction after turning, i forgot that the direction can max out in the anticlockwise direction too, but just set it to north when a critical point in both clockwise and anticlockwise cases is reached...
Part2 problem, I actually felt that it was not much of a problem, and flexed my slow brain in order to get efficient code. Made static classes for the ship and waypoint, so can seperate between them easier, and i am proud of my Waypoint Rotate method, which readjusts the relative east and relative north of the waypoint, by setting its negative counter parts.
code:

    if (clockwise)
            (RelativeEast, RelativeNorth) = (RelativeNorth, -RelativeEast);
    else
            (RelativeEast, RelativeNorth) = (-RelativeNorth, RelativeEast);