Pushing day10
This challenge was a disaster...
Part1: 27 mins
Part2: 2 days....

Honestly, I didn't know how to count the number of possible arrangements at all.
Tried all sorts of wacky methods.
It came down to actually writing down the possible things on paper.
The most unexpected was to factorise the number is the test data answers, which came out ot be the answer...